refactor(manifest): move comparison DTOs to ComparisonModels.cs (#601 PR1)
Extract the 15 public data types (LoadedManifest and its nested config
shapes, ComparisonResult/SummaryResult/BatesAnalysis/BatesRangeReport/
VolumeResult/ResultDetails, and the RecordDetail/ChangedDetail/
ReplacedDetail/DuplicateDetail detail records) from the bottom of
ProductionManifestComparer.cs into a dedicated ComparisonModels.cs file
in the same Zipper.ManifestComparison namespace.

Pure file move: no behavior, no namespace, no visibility, and no JSON
contract changes. The comparer retains its System.Text.Json.Serialization
using for JsonIgnoreCondition. ComparisonTests and
ProductionManifestComparerTests still pass (29/29), preserving output
parity.

First slice of the #601 god-class decomposition (DTOs only); subsequent
PRs extract the loader, comparison engine, and report writer.
